requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.109.0
ffmpeg-python==0.2.0
future==0.18.3
h11==0.14.0
idna==3.6
joblib==1.3.2
nltk==3.8.1
numpy==1.26.3
pandas==2.1.4
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.4.0
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
SpeechRecognition==3.10.1
starlette==0.35.1
threadpoolctl==3.2.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.26.0
